Remove the requires clause from the M6502 template
These two requirements mean class B cannot contain a pointer to
M6502<B> because B is an incomplete type at the time the pointer
is defined. Since they are just there to reduce the number of
error messages if B is missing readAddr() and writeAddr(), removing
them has no impact on the actual functionality of M6502.

One possible way I can think of to avoid this without removing the
requires clause is by creating a subclass of B to hold the pointer
and then only using the subclass, but that feels dumb to do unless
it's absolutely necessary, which it really isn't.