Added support for a break handler
If the bus used to instantiate M6502 has a breakHandler() function, that
function will be called every time a BRK instruction is fetched. It
receives a reference to the CPU that fetched it and its address. PC is
positioned at the byte immediately following it. If the handler returns
true, then the standard BRK sequence is skipped and execution continues
at PC. (Note that the PC will not skip the byte after the BRK instruction
in this case.) If the handler returns false, then the BRK sequence will
continue as normal.