Added Wolfgang Lorenz's C64 Test Suite
This is the slightly modified version of the suite from VICE because
I needed a version that would read the "magic constant" for the
unstable opcodes instead of assuming one. The test suite passes through
the cputiming tests. Getting further requires a more accurate CIA
emulation, which may or may not be worthwhile. The irq and nmi tests
sound like they might be useful, but I think the rest are unnecessary
for testing just a 6502.