Modify the branchwrap test to use the VICE framework
When loaded from trap17, branchwrap failed because its entry point was
not at the location the VICE framework expected and it, so execution
started inside part of the SYS statement.

Since I was modifying it anyway, it now skips the MMU tests and goes
straight to the cputiming test, so I can remove special handling for its
LOAD call.